Refactor agent configuration files and update import paths
