+++ suites/agents/agents.test.ts
@@ -4,10 +4,10 @@ import { logError } from "@helpers/logger";
 import { sendAgentNotification } from "@helpers/notifications";
 import { verifyBotMessageStream } from "@helpers/streams";
 import { setupTestLifecycle } from "@helpers/vitest";
-import productionAgents from "@inboxes/agents.json";
+import productionAgents from "./agents.json";
 import { typeOfResponse, typeofStream, typeOfSync } from "@workers/main";
 import { getWorkers, type WorkerManager } from "@workers/manager";
-import { type AgentConfig } from "../../types/agents";
+import { type AgentConfig } from "./agents";
 import { IdentifierKind, type Dm } from "@xmtp/node-sdk";
 import { beforeAll, describe, expect, it } from "vitest";
